Moved confiy.py to ParametricEQconfig.py to prevent name conflicts
