+++ Equalizer.py
@@ -16,7 +16,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>
 
 from gi.repository import Gtk
-from config import Config
+from ParametricEQconfig import Config
 from EQBandParams import EQBandParams
 #import os, sys, inspect
