+++ ParametricEQconfig.py
